Fix Password issue in entryMenu Add DataGrid In تسجيل أنواع الشركات --> Please Review ya Bombo
