+++ CCL Oil Labs Control/Views/EditUsrMenu.xaml
@@ -1,4 +1,5 @@
-﻿<UserControl x:Class="CCL_Oil_Labs_Control.Views.EditUsrMenu"
+﻿
+    <UserControl x:Class="CCL_Oil_Labs_Control.Views.EditUsrMenu"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
